backend/requirements.txt
# Core Flask Dependencies
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-JWT-Extended==4.6.0
Flask-CORS==4.0.0

# Supabase & Database
supabase==2.8.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0

# Utility Libraries
qrcode[pil]==7.4.2
Pillow==10.1.0
requests==2.31.0
python-dateutil==2.8.2

# Development & Testing
pytest==7.4.3
pytest-flask==1.3.0

# Production Dependencies
gunicorn==21.2.0
redis==5.0.1

# Existing dependencies (if any)
werkzeug==3.0.1 
 